+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.0.1
 
 MVP nodes:
+- `Add String to Format-Dict`
 - `Format-Dict from Text`
 - `Format-Dict Preview`
 - `String Formatter`